Tinder for dogs
TinDog is a front-end website for dog's owners who are willing to cross their pet with other.A tinder website made exclusively for Dogs.The main framework used in building the TinDog website is Bootstrap and its components like Carousel, Card, Buttons, Navbar etc. It also used HTML5, CSS3 and Google Fonts and Font Awesome.A home page website which provides details of the application to the end user like the desscription of the website, from where we can download the application, a section that displays the feedback from the clients and different plans and subscriptions.